Started working on the Admin
I realised the current folder setup will not work. I thought it would be smart to have admin as its own app, but this seems to cause trouble. I have not deleted it yet, but I might do it soon.

I created the model and the admin part to add FAQ questions, and made a basic view that could display it. The design is quite terrible, but the funcitonality works fine...