I have made some structural minor changes.
I have begun working on the student section, and are trying to figure out how to create sign in capabilities. It seems to be rather easy, since there is no tutorials on the subject. I will just have to read through the Django Documentation, to figure it out.

Until then I have created views in both the company section and the student section, just to create the same structure in root.

I have a theory that I will enable login restriction in urls.py in the leapkit section. This way there will be no need to think about restrictions in each section.