Renamed settings files and included some forms
I renamed the settings files to make their purpose clearer. On top of this I removed some static bindings in the different apps, after reading this was an antipattern

I included some forms for logging in and signing up, but have not implemented their funcitionality yet, since I still don't know how to work with users. I included a JavaScript library for making it easy to validate the form, and give the user constructive feedback.

I have not focused on the design more than the absolute minimum, since this might be the job of someone else. and the flow might need to be changed. At this step it is mostly just to get the functionality in place.