Finally got it to work!
After countless of hours, a mild depression, a some tears and a few outcries, the Contact form is implemented..

Now the contact form is working as is should. It is still not tested, but I will focus on some other parts from now on, since I am sick and tired of working on it.

I have started working with PyCharm, and it is a wonderfull tool.. I don't think I will go back.. It refactor the code to comply with the pep8.. So that is cool..