Started working on Students
I have now begun working on model for the students. I have created the foundation for the forms used to create and change the basic information as well.

I have not tested the code yet, and I have not tried to implement to forms. Much more work is needed, but I have to stop for today.