Basic user functionality implemented
I have now implemented basic Company and Student user functionality. To go to either section of the website, the user has to sign in. The user can sign in from the same page and is then directed to the student section or the company section based on the group to which he belongs.

If the user does not have a profile they can register as either company or Student. The registration form is the same, except the group they are put in.

Ideally I would like to have two different kinds of users, but so far I have not found a way to implement that.