Custom user model implemented
I have now created a custom user model, to make it easier to work with users in the future. The user now authenticates using their Email instead of a chosen identifier.

In the process I feel I have learned a lot about the idea of a user in django, so tomorrow I will work with extending the usermodel so I can make student users and company users. When that is done, I will start testing and work with south.. Every time I make changes to the model I delete the db and create it again. This is time consuming and stupid. I will make sure to understand south, so I do not have to do this anymore.