Refactored, Renamed and got student sign up / sign in working
I had a lot of problems with proper signup. I tried several different ways, in an attempt to find the smoothest sign up process. The problem I had was to create a profile for the user depending on the user being a student or a company. I tried to do this at the same time as sign up, but this turned out to be more problematic than suspected.

Now the user will have to sign in, and then be directly redirected to the student profile creation page. This is okay, but not optimal.

I also had some clashes with my naming of the users classes. My naming interfeared with some of the apps I was using, and I had to renamed users to custom user and the app to leapkit_users.

I had some problems with the backens I had implemented, and which destroyed all sign in functionality across the site.

Today I will focus on testing.. Every time a problem occurs I have to test several different things over and over again, and it is truly annoying..